qa-project/code/requirements.txt
pytest==5.4.2
pytest-xdist==1.32.0
requests==2.23.0
selenium==3.141.0
webdriver-manager==2.5.2
SQLAlchemy==1.3.17
Faker==4.1.0
Flask==1.1.2
PyMySQL==0.9.3
allure-pytest==2.8.16
